refactor: clarify domain and platform boundaries
