+++ src/views/Advanced.tsx
@@ -3,8 +3,7 @@ import { type FC, useState } from "react";
 
 import { SaveWithPreview, ToggleFilePicker } from "../components";
 import { useOptions } from "../hooks";
-import { browseFiles } from "../infra/client";
-import { getHomePath } from "../infra/environment";
+import { browseFiles, getHomePath } from "../infra/decky";
 import { t } from "../utils/translate";
 
 const Advanced: FC = () => {
