+++ tests/launchOptions.test.ts
@@ -2,14 +2,14 @@ import { describe, expect, it } from "vitest";
 
 import { type LaunchOption, LaunchOptions, type LaunchOptionsEditResult } from "../src/domain/launchOptions";
 
-const editedValue = (result: LaunchOptionsEditResult): LaunchOptions => {
+const expectSuccess = (result: LaunchOptionsEditResult): LaunchOptions => {
   expect(result.ok).toBe(true);
   if (!result.ok) throw new Error(result.error);
   return result.value;
 };
 
 describe("LaunchOptions", () => {
-  describe("parsing and preservation", () => {
+  describe("reading and preservation", () => {
     it("preserves the source exactly and exposes business getters", () => {
       const source =
         '  DXVK_ASYNC=1\tRADV_PERFTEST=gpl LANG="ja_JP.utf8" HOST_LC_ALL=ja_JP.utf8 ' +
@@ -34,7 +34,7 @@ describe("LaunchOptions", () => {
 
     it("preserves unknown parameters, whitespace, and quoting while patching one token", () => {
       const source = '  MANGOHUD="1"\t%command%   --mystery "a b"  ';
-      const edited = editedValue(LaunchOptions.parse(source).setDxvkAsync(true));
+      const edited = expectSuccess(LaunchOptions.parse(source).setDxvkAsync(true));
 
       expect(edited.toString()).toContain('  MANGOHUD="1"\t');
       expect(edited.toString()).toContain('%command%   --mystery "a b"  ');
@@ -43,7 +43,7 @@ describe("LaunchOptions", () => {
 
     it("does not mutate the original instance", () => {
       const original = LaunchOptions.parse("FOO=1 %command%");
-      const edited = editedValue(original.setLanguage("fr_FR.utf8"));
+      const edited = expectSuccess(original.setLanguage("fr_FR.utf8"));
 
       expect(original.toString()).toBe("FOO=1 %command%");
       expect(original.language).toBeUndefined();
@@ -57,15 +57,15 @@ describe("LaunchOptions", () => {
     });
   });
 
-  describe("business edits", () => {
+  describe("managed options", () => {
     it("sets and disables trainer options", () => {
-      const enabled = editedValue(LaunchOptions.parse("").setTrainer("/home/deck/trainers/game.exe"));
+      const enabled = expectSuccess(LaunchOptions.parse("").setTrainer("/home/deck/trainers/game.exe"));
 
       expect(enabled.trainerPath).toBe("/home/deck/trainers/game.exe");
       expect(enabled.trainerDirectory).toBe("/home/deck/trainers");
       expect(enabled.isTrainerEnabled).toBe(true);
 
-      const disabled = editedValue(enabled.disableTrainer());
+      const disabled = expectSuccess(enabled.disableTrainer());
       expect(disabled.isTrainerEnabled).toBe(false);
       expect(disabled.trainerPath).toBeUndefined();
       expect(disabled.trainerDirectory).toBeUndefined();
@@ -73,40 +73,40 @@ describe("LaunchOptions", () => {
     });
 
     it("sets and disables both language variables", () => {
-      const enabled = editedValue(LaunchOptions.parse("%command%").setLanguage("de_DE.UTF-8"));
+      const enabled = expectSuccess(LaunchOptions.parse("%command%").setLanguage("de_DE.UTF-8"));
       expect(enabled.language).toBe("de_DE.UTF-8");
       expect(enabled.isLanguageEnabled).toBe(true);
       expect(enabled.toString()).toContain("LANG=de_DE.UTF-8");
       expect(enabled.toString()).toContain("HOST_LC_ALL=de_DE.UTF-8");
 
-      const disabled = editedValue(enabled.disableLanguage());
+      const disabled = expectSuccess(enabled.disableLanguage());
       expect(disabled.isLanguageEnabled).toBe(false);
       expect(disabled.toString()).toBe("");
     });
 
     it("sets and disables a compatibility path", () => {
-      const enabled = editedValue(
+      const enabled = expectSuccess(
         LaunchOptions.parse("UNKNOWN=keep %command%").setCompatibilityPath("/tmp/prefix one"),
       );
       expect(enabled.compatibilityPath).toBe("/tmp/prefix one");
 
-      const disabled = editedValue(enabled.disableCompatibilityPath());
+      const disabled = expectSuccess(enabled.disableCompatibilityPath());
       expect(disabled.compatibilityPath).toBeUndefined();
       expect(disabled.toString()).toMatch(/UNKNOWN=keep\s+%command%/);
     });
 
     it("toggles DXVK, RADV, and Lossless Scaling", () => {
       let options = LaunchOptions.parse("");
-      options = editedValue(options.setDxvkAsync(true));
-      options = editedValue(options.setRadvPerftest(true));
-      options = editedValue(options.setLosslessScaling(true));
+      options = expectSuccess(options.setDxvkAsync(true));
+      options = expectSuccess(options.setRadvPerftest(true));
+      options = expectSuccess(options.setLosslessScaling(true));
       expect(options.isDxvkAsyncEnabled).toBe(true);
       expect(options.isRadvPerftestEnabled).toBe(true);
       expect(options.isLosslessScalingEnabled).toBe(true);
 
-      options = editedValue(options.setDxvkAsync(false));
-      options = editedValue(options.setRadvPerftest(false));
-      options = editedValue(options.setLosslessScaling(false));
+      options = expectSuccess(options.setDxvkAsync(false));
+      options = expectSuccess(options.setRadvPerftest(false));
+      options = expectSuccess(options.setLosslessScaling(false));
       expect(options.isDxvkAsyncEnabled).toBe(false);
       expect(options.isRadvPerftestEnabled).toBe(false);
       expect(options.isLosslessScalingEnabled).toBe(false);
@@ -114,34 +114,34 @@ describe("LaunchOptions", () => {
     });
 
     it("keeps Framegen patch and unpatch mutually exclusive", () => {
-      let options = editedValue(LaunchOptions.parse("").setFramegenPatch(true));
+      let options = expectSuccess(LaunchOptions.parse("").setFramegenPatch(true));
       expect(options.isFramegenPatchEnabled).toBe(true);
       expect(options.isFramegenUnpatchEnabled).toBe(false);
 
-      options = editedValue(options.setFramegenUnpatch(true));
+      options = expectSuccess(options.setFramegenUnpatch(true));
       expect(options.isFramegenPatchEnabled).toBe(false);
       expect(options.isFramegenUnpatchEnabled).toBe(true);
 
-      options = editedValue(options.setFramegenPatch(true));
+      options = expectSuccess(options.setFramegenPatch(true));
       expect(options.isFramegenPatchEnabled).toBe(true);
       expect(options.isFramegenUnpatchEnabled).toBe(false);
     });
 
     it("removes duplicate managed entries and leaves unknown entries intact", () => {
       const source = "DXVK_ASYNC=1 FOO=keep DXVK_ASYNC=0 DXVK_ASYNC=1 %command%";
-      const enabled = editedValue(LaunchOptions.parse(source).setDxvkAsync(true));
+      const enabled = expectSuccess(LaunchOptions.parse(source).setDxvkAsync(true));
       expect(enabled.toString().match(/DXVK_ASYNC=/g)).toHaveLength(1);
       expect(enabled.toString()).toContain("FOO=keep");
 
-      const disabled = editedValue(enabled.setDxvkAsync(false));
+      const disabled = expectSuccess(enabled.setDxvkAsync(false));
       expect(disabled.toString()).not.toContain("DXVK_ASYNC=");
       expect(disabled.toString()).toContain("FOO=keep");
     });
 
     it("supports consecutive immutable edits", () => {
-      const first = editedValue(LaunchOptions.parse("").setLanguage("en_US.UTF-8"));
-      const second = editedValue(first.setDxvkAsync(true));
-      const third = editedValue(second.setCompatibilityPath("/home/deck/prefix"));
+      const first = expectSuccess(LaunchOptions.parse("").setLanguage("en_US.UTF-8"));
+      const second = expectSuccess(first.setDxvkAsync(true));
+      const third = expectSuccess(second.setCompatibilityPath("/home/deck/prefix"));
 
       expect(first.isDxvkAsyncEnabled).toBe(false);
       expect(second.compatibilityPath).toBeUndefined();
@@ -172,18 +172,18 @@ describe("LaunchOptions", () => {
       ];
       let options = LaunchOptions.parse("");
 
-      for (const option of customOptions) options = editedValue(options.setCustomOption(option, true));
+      for (const option of customOptions) options = expectSuccess(options.setCustomOption(option, true));
       for (const option of customOptions) expect(options.isCustomOptionEnabled(option)).toBe(true);
 
-      for (const option of customOptions) options = editedValue(options.setCustomOption(option, false));
+      for (const option of customOptions) options = expectSuccess(options.setCustomOption(option, false));
       for (const option of customOptions) expect(options.isCustomOptionEnabled(option)).toBe(false);
       expect(options.toString()).toBe("");
     });
 
     it("matches and deduplicates a quoted prefix command", () => {
       const option: LaunchOption = { type: "pre_cmd", key: 'echo "a b"' };
-      const enabled = editedValue(LaunchOptions.parse("").setCustomOption(option, true));
-      const enabledAgain = editedValue(enabled.setCustomOption(option, true));
+      const enabled = expectSuccess(LaunchOptions.parse("").setCustomOption(option, true));
+      const enabledAgain = expectSuccess(enabled.setCustomOption(option, true));
 
       expect(enabled.isCustomOptionEnabled(option)).toBe(true);
       expect(enabledAgain.isCustomOptionEnabled(option)).toBe(true);
@@ -192,17 +192,17 @@ describe("LaunchOptions", () => {
 
     it("does not remove a valued option when disabling a no-value option with the same key", () => {
       const source = "%command% -width 1920";
-      const edited = editedValue(
+      const edited = expectSuccess(
         LaunchOptions.parse(source).setCustomOption({ type: "flag_args", key: "-width" }, false),
       );
       expect(edited.toString()).toBe(source);
     });
 
     it("replaces an existing value with the same type and key when enabling", () => {
-      const env = editedValue(
+      const env = expectSuccess(
         LaunchOptions.parse("FOO=old %command%").setCustomOption({ type: "env", key: "FOO", value: "new" }, true),
       );
-      const flag = editedValue(
+      const flag = expectSuccess(
         LaunchOptions.parse("%command% -width 1280").setCustomOption(
           { type: "flag_args", key: "-width", value: "1920" },
           true,
@@ -232,7 +232,7 @@ describe("LaunchOptions", () => {
     });
   });
 
-  describe("markers and blank input", () => {
+  describe("edit validation", () => {
     it.each(["FOO=1", "FOO=1 %command% %command%"])("rejects edits without exactly one marker: %s", (source) => {
       const original = LaunchOptions.parse(source);
       const result = original.setDxvkAsync(true);
@@ -241,24 +241,24 @@ describe("LaunchOptions", () => {
     });
 
     it.each(["", "   \t "])("allows edits of blank input: %j", (source) => {
-      const edited = editedValue(LaunchOptions.parse(source).setDxvkAsync(true));
+      const edited = expectSuccess(LaunchOptions.parse(source).setDxvkAsync(true));
       expect(edited.isDxvkAsyncEnabled).toBe(true);
       expect(edited.toString()).toContain("%command%");
     });
 
     it("supports enabling and then disabling an initially empty configuration", () => {
-      const enabled = editedValue(LaunchOptions.parse("").setDxvkAsync(true));
-      const disabled = editedValue(enabled.setDxvkAsync(false));
+      const enabled = expectSuccess(LaunchOptions.parse("").setDxvkAsync(true));
+      const disabled = expectSuccess(enabled.setDxvkAsync(false));
       expect(disabled.toString()).toBe("");
     });
   });
 
-  describe("quoting and token boundaries", () => {
+  describe("source-safe editing", () => {
     it("round-trips trainer and compatibility paths with shell-special characters", () => {
       const directory = '/home/deck/C:\\Games\\it\'s "odd" $cash `tick`';
       const path = `${directory}/trainer.exe`;
-      let options = editedValue(LaunchOptions.parse("").setTrainer(path));
-      options = editedValue(options.setCompatibilityPath(directory));
+      let options = expectSuccess(LaunchOptions.parse("").setTrainer(path));
+      options = expectSuccess(options.setCompatibilityPath(directory));
 
       expect(options.trainerPath).toBe(path);
       expect(options.trainerDirectory).toBe(directory);
@@ -271,7 +271,7 @@ describe("LaunchOptions", () => {
     });
 
     it("does not delete environment variables between prefix commands", () => {
-      const edited = editedValue(
+      const edited = expectSuccess(
         LaunchOptions.parse("cmd1 FOO=1 cmd2 %command%").setCustomOption({ type: "pre_cmd", key: "cmd1" }, false),
       );
       expect(edited.toString()).not.toContain("cmd1");
@@ -279,10 +279,10 @@ describe("LaunchOptions", () => {
     });
 
     it("removes the correct separator when deleting either prefix command", () => {
-      const firstRemoved = editedValue(
+      const firstRemoved = expectSuccess(
         LaunchOptions.parse("cmd1 -- cmd2 %command%").setCustomOption({ type: "pre_cmd", key: "cmd1" }, false),
       );
-      const secondRemoved = editedValue(
+      const secondRemoved = expectSuccess(
         LaunchOptions.parse("cmd1 -- cmd2 %command%").setCustomOption({ type: "pre_cmd", key: "cmd2" }, false),
       );
 
@@ -294,32 +294,10 @@ describe("LaunchOptions", () => {
 
     it("removes duplicate prefix commands without leaving a separator", () => {
       const option: LaunchOption = { type: "pre_cmd", key: "cmd" };
-      const edited = editedValue(LaunchOptions.parse("cmd -- cmd -- cmd %command%").setCustomOption(option, true));
+      const edited = expectSuccess(LaunchOptions.parse("cmd -- cmd -- cmd %command%").setCustomOption(option, true));
 
       expect(edited.toString().match(/cmd/g)).toHaveLength(1);
       expect(edited.toString()).not.toContain("--");
-    });
-
-    it("keeps prefix arguments containing equals signs out of environment entries", () => {
-      const option: LaunchOption = { type: "pre_cmd", key: "gamescope --foo=bar" };
-      const enabled = editedValue(LaunchOptions.parse("").setCustomOption(option, true));
-
-      expect(enabled.isCustomOptionEnabled(option)).toBe(true);
-    });
-
-    it("treats a quoted leading-hyphen token as a flag value", () => {
-      const options = LaunchOptions.parse('%command% -x "-foo"');
-
-      expect(options.isCustomOptionEnabled({ type: "flag_args", key: "-x", value: "-foo" })).toBe(true);
-      expect(options.isCustomOptionEnabled({ type: "flag_args", key: "-foo" })).toBe(false);
-    });
-
-    it("parses a negative flag argument as its exact value", () => {
-      const options = LaunchOptions.parse("%command% -width -1 -height -2.5 -novid");
-      expect(options.isCustomOptionEnabled({ type: "flag_args", key: "-width", value: "-1" })).toBe(true);
-      expect(options.isCustomOptionEnabled({ type: "flag_args", key: "-height", value: "-2.5" })).toBe(true);
-      expect(options.isCustomOptionEnabled({ type: "flag_args", key: "-width" })).toBe(false);
-      expect(options.isCustomOptionEnabled({ type: "flag_args", key: "-novid" })).toBe(true);
     });
   });
 });
